requirements.txt
opencv-python
pillow
easyocr
numpy
tensorflow==2.4.0
matplotlib
